move e2e test
